Fix Swiftlint
