+++ Allergy/AllergyAppDelegate.swift
@@ -15,7 +15,6 @@ import FHIRToFirestoreAdapter
 import FirebaseAccount
 import class FirebaseFirestore.FirestoreSettings
 import FirebaseStorage
-import FirebaseConfiguration
 import FirestoreDataStorage
 import FirestoreStoragePrefixUserIdAdapter
 import HealthKit
@@ -77,16 +76,3 @@ class AllergyAppDelegate: CardinalKitAppDelegate {
         }
     }
 }
-
-
-class FirestoreStorage<ComponentStandard: Standard>: Component {
-     @Dependency private var configureFirebaseApp: ConfigureFirebaseApp
-
-
-     func configure() {
-         let storage = Storage.storage()
-         if FeatureFlags.useFirebaseEmulator {
-             storage.useEmulator(withHost: "localhost", port: 9199)
-         }
-     }
- }
